+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Unreleased` header.
   misinterpreted during a drag and drop operation.
 - On X11, fix a bug where focusing the window would panic.
 - On macOS, fix `refresh_rate_millihertz`.
+- On Windows, add `with_system_backdrop`.
 
 # 0.29.4